Add parse option to display points-to set information for pointer dereference
The points to set information provides some intuition about the complexity
of the resulting case split expression in case of a pointer dereference.
This commit adds parse option --show-points-to-sets to track the size and
contents of the points-to set for pointers.

cpp lint bmc_util.h

Add error message for plain and xml format output